bankapp-core: implemented mongodb repository

+++ bankapp/src/main/kotlin/com/clouway/bankapp/core/SessionCounter.kt
@@ -4,5 +4,5 @@ package com.clouway.bankapp.core
  * @author Tsvetozar Bonev (tsbonev@gmail.com)
  */
 interface SessionCounter {
-    fun getActiveSessionsCount(): Int
+    fun getActiveSessionsCount(): Long
 }
